LayoutTests/platform/ios-sim/storage/indexeddb/mozilla/object-store-remove-values-expected.txt

CONSOLE MESSAGE: line 190: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
Test IndexedDB: adds, gets, deletes, and re-gets a record in a variety of different object store configurations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


data = [
        { name: 'inline key; key generator',
          autoIncrement: true,
          storedObject: {name: 'Lincoln'},
          keyName: 'id',
          keyValue: undefined,
        },
        { name: 'inline key; no key generator',
          autoIncrement: false,
          storedObject: {id: 1, name: 'Lincoln'},
          keyName: 'id',
          keyValue: undefined,
        },
        { name: 'out of line key; key generator',
          autoIncrement: true,
          storedObject: {name: 'Lincoln'},
          keyName: undefined,
          keyValue: undefined,
        },
        { name: 'out of line key; no key generator',
          autoIncrement: false,
          storedObject: {name: 'Lincoln'},
          keyName: undefined,
          keyValue: 1,
        }
    ];
i = 0;

indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "object-store-remove-values.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
